+++ src/lib/ast/minify.ts
@@ -1,5 +1,5 @@
 import { eq } from "../parser/utils/eq.ts";
-import { doRender, renderToken } from "../renderer/render.ts";
+import { doRender, renderValue } from "../renderer/render.ts";
 import * as allFeatures from "./features/index.ts";
 import { walkValues } from "./walk.ts";
 import type {
@@ -398,7 +398,7 @@ function minifyAtRuleMedia(tokens: Token[]): Token[] {
                 }
 
                 acc.set(
-                    values.reduce((acc, t) => acc + renderToken(t), ""),
+                    values.reduce((acc, t) => acc + renderValue(t), ""),
                     result.values,
                 );
 
@@ -592,7 +592,7 @@ function doMinify(
                                         (index + 3 < arr.length ||
                                             arr[index + 2]?.typ === EnumToken.WhitespaceTokenType))
                                         ? ""
-                                        : renderToken(curr)),
+                                        : renderValue(curr)),
                                 "",
                             );
                         }
@@ -630,7 +630,7 @@ function doMinify(
                                     arr[index + 1]?.typ === EnumToken.CommentTokenType &&
                                     (index + 3 < arr.length || arr[index + 2].typ === EnumToken.WhitespaceTokenType))
                                     ? ""
-                                    : renderToken(curr)),
+                                    : renderValue(curr)),
                             "",
                         );
                     }
@@ -1508,7 +1508,7 @@ function fixSelector(node: AstRule): void {
             }
         }
 
-        node.sel = attributes.reduce((acc: string, curr: Token) => acc + renderToken(curr), "");
+        node.sel = attributes.reduce((acc: string, curr: Token) => acc + renderValue(curr), "");
     }
 }